# Conflicts:
#	common/src/main/java/com/jjg/game/common/utils/TimeHelper.java

+++ hall/src/main/java/com/jjg/game/hall/friendroom/services/FriendRoomServices.java
@@ -123,7 +123,7 @@ public class FriendRoomServices {
         if (playerPack == null && !useGold) {
             return Code.NOT_ENOUGH;
         } else if (playerPack != null) {
-            CommonResult<Integer> removeRes = playerPack.removeItem(itemId, reqCreateFriendsRoom.itemNum);
+            CommonResult<Long> removeRes = playerPack.removeItem(itemId, reqCreateFriendsRoom.itemNum);
             if (!removeRes.success()) {
                 return removeRes.code;
             }
